fix: make tests build and some refactoring
